Fix llama3 when torch compile with capturing scalar outputs
Without this workaround, it will throw an error: "Could not extract specialized
integer from data-dependent expression"

This is caused by the `torch.searchsorted` in `llama3_flash_attn_prepare_cu_seqlens`,
the total length is dependent on the last element in `cu_seqlens`. If we
convert the tensor to scalar with `.item()` here, it will create a situation
similar to https://github.com/pytorch/pytorch/issues/150613

This workaround resolves this by delaying the tensor to scalar conversion after
`torch.searchsorted` finishes. This is OK as the second param `values` of
`torch.searchsorted` also accept the tensor type.
https://pytorch.org/docs/stable/generated/torch.searchsorted.html

Signed-off-by: Hollow Man <hollowman@opensuse.org>